+++ assets/javascripts/kitten/components/buttons/button-group/test.js
@@ -12,7 +12,7 @@ describe('<ButtonGroup />', () => {
         .create(
           <ButtonGroup>
             <ButtonGroup.Button>Button1</ButtonGroup.Button>
-            <ButtonGroup.Button>Button2</ButtonGroup.Button>
+            <ButtonGroup.Button active>Button2</ButtonGroup.Button>
             <ButtonGroup.Button>Button3</ButtonGroup.Button>
           </ButtonGroup>,
         )