+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Fix: Fix `LinkBox` by using direct props instead of `linkProps`.
+  `linkProps` is now deprecated
+
 ## [2.75.0] - 2020-06-15
 
 Doc: